+++ prac04/player_bullet_spread.h
@@ -2,12 +2,12 @@
 #include "Bullet.h"
 class PlayerBulletSpread : public Bullet {
 public:
-	PlayerBulletSpread(int x, int y, float velX, float velY);
+	PlayerBulletSpread(float x, float y, float velX, float velY);
 	void Update() override;
 	void Render() override;
 
-	int posX; 
-	int posY;
+	float posX; 
+	float posY;
 
 	float velX;
 	float velY;
